+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchInit.kt
@@ -12,9 +12,10 @@ class OverdueSearchInit(
     val effects = mutableSetOf<OverdueSearchEffect>()
     if (isOverdueSearchV2Enabled) {
       effects.add(LoadVillageAndPatientNames)
+    } else {
+      effects.add(LoadOverdueSearchHistory)
     }
 
-    effects.add(LoadOverdueSearchHistory)
     effects.add(LoadSelectedOverdueAppointmentIds)
 
     return first(model, effects)