+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Bump Sentry gradle plugin to v3.1.3
 - Bump dagger to v2.43
 - Bump mixpanel to v6.5.1
+- Add overdue search v2 feature flag
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-25-8342