+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Display search suggestions after loading in overdue search screen
 - Load village and patient names when overdue search screen is created
 - Don't show search history when overdue search v2 is enabled
+- Bump room to 2.4.3
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-25-8342